Use our own "version" of setState
Given that we already work around the fact that React's .setState() is not
synchronous with respect to .state, we might as well do away with using
React's setState altogether and use forceUpdate. It's much cleaner and
we don't have to worry at all then about playing nicely with React
upgrades.

Similar in approach to https://blog.cloudboost.io/3-reasons-why-i-stopped-using-react-setstate-ab73fc67a42e